do not use explicitysorter in conf
