+++ docs/source/conf.py
@@ -85,23 +85,23 @@ sphinx_gallery_conf = {
     'matplotlib_animations': True,
     'image_srcset': ["2x"],
 
-    'within_subsection_order': ExplicitOrder([
-        '../../examples/plot',
-        '../../examples/scatter',
-        '../../examples/imshow',
-        '../../examples/bar_chart',
-        '../../examples/hist',
-        '../../examples/lollipop_plot',
-        '../../examples/dumbell_plot',
-        '../../examples/regplot',
-        '../../examples/ridge',
-        '../../examples/pie',
-        '../../examples/contour',
-        '../../examples/circular_bar_plot',
-        '../../examples/parallel_coordinates',
-        '../../examples/spider_plot',
-        '../../examples/taylor_plot',
-                                       ]),
+    'within_subsection_order':  ExampleTitleSortKey,
+        # '../../examples/plot',
+        # '../../examples/scatter',
+        # '../../examples/imshow',
+        # '../../examples/bar_chart',
+        # '../../examples/hist',
+        # '../../examples/lollipop_plot',
+        # '../../examples/dumbell_plot',
+        # '../../examples/regplot',
+        # '../../examples/ridge',
+        # '../../examples/pie',
+        # '../../examples/contour',
+        # '../../examples/circular_bar_plot',
+        # '../../examples/parallel_coordinates',
+        # '../../examples/spider_plot',
+        # '../../examples/taylor_plot',
+#                                       ]),
 }
 
 # -- Options for HTML output -------------------------------------------------
