+++ tests/test_main.py
@@ -17,7 +17,6 @@ from easy_mpl import regplot, scatter, contour
 from easy_mpl.utils import BAR_CMAPS, make_cols_from_cmap
 from easy_mpl import dumbbell_plot, ridge
 from easy_mpl import parallel_coordinates
-from easy_mpl import lollipop_plot
 
 
 def get_chart_data(n):
@@ -775,56 +774,6 @@ class TestParallelCoord(unittest.TestCase):
                                   categories=categories,
                                   title="nan in data and categories",
                                   show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-
-class TestLollipopPlot(unittest.TestCase):
-    show = False
-
-    y = np.random.randint(0, 10, size=10)
-
-    def test_vanilla(self):
-        ax = lollipop_plot(self.y, title="vanilla", show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_with_x_and_y(self):
-        ax = lollipop_plot(self.y, np.linspace(0, 100, len(self.y)),
-         title="with x and y", show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_custom_linestyle(self):
-
-        ax = lollipop_plot(self.y, line_style='--', title="with custom linestyle",
-         show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_custom_marker_style(self):
-
-        ax = lollipop_plot(self.y, marker_style='D', title="with custom marker style",
-         show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_sort(self):
-
-        ax = lollipop_plot(self.y, title="sort", sort=True, show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_horizontal(self):
-        y = np.random.randint(0, 20, size=10)
-        ax = lollipop_plot(y, title="horizontal", orientation="horizontal", show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_with_nan_vals(self):
-        y = np.random.randint(0, 20, size=10).astype("float32")
-        y[3] = np.nan
-        ax = lollipop_plot(y, title="with nan vals", show=self.show)
         assert isinstance(ax, plt.Axes)
         return
